ch: refactor adding new case functionality
